refactor: move ooni into the internal package
With this commit we've reorganised the package structure. Now I will
go more in depth and perform further refactoring.
