+++ internal/cli/root/root.go
@@ -3,7 +3,7 @@ package root
 import (
 	"github.com/alecthomas/kingpin"
 	"github.com/apex/log"
-	ooni "github.com/ooni/probe-cli"
+	"github.com/ooni/probe-cli/internal/ooni"
 	"github.com/ooni/probe-cli/internal/log/handlers/batch"
 	"github.com/ooni/probe-cli/internal/log/handlers/cli"
 	"github.com/ooni/probe-cli/internal/utils"
