+++ internal/nettests/nettests.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/apex/log"
 	"github.com/fatih/color"
-	ooni "github.com/ooni/probe-cli"
+	"github.com/ooni/probe-cli/internal/ooni"
 	"github.com/ooni/probe-cli/internal/database"
 	"github.com/ooni/probe-cli/internal/output"
 	engine "github.com/ooni/probe-engine"
